requirements.txt
# Core
pydantic-settings
python-dotenv
aiohttp
telethon

# AI & NLP
openai
langchain
langchain-openai
langchain-community
sentence-transformers
transformers
nltk

# Database
sqlmodel
chromadb
psycopg2-binary
pgvector

# Utilities
colorama
rich
questionary